requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.2
django-mail-templated==2.6.5
django-sendgrid-v5==0.9.0
future==0.18.2
gunicorn==20.1.0
Pillow==8.0.1
psycopg2==2.9.1
python-http-client==3.3.1
pytz==2020.1
sendgrid==6.4.7
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
whitenoise==5.3.0
